+++ Gen3/Frame3.cpp
@@ -35,21 +35,21 @@ Frame3::Frame3(u16 tid, u16 sid, u16 psv)
 
 QString Frame3::getTime()
 {
-    int32_t seconds = frame / 60;
-    int32_t milliseconds = ((frame % 60) * 100) / 60;
+    int seconds = frame / 60;
+    int milliseconds = ((frame % 60) * 100) / 60;
     if (seconds < 60)
     {
         return QString("%1.%2 s").arg(seconds).arg(milliseconds, 2, 10, QChar('0'));
     }
 
-    int32_t minutes = frame / 3600;
+    int minutes = frame / 3600;
     seconds -= minutes * 60;
     if (minutes < 60)
     {
         return QString("%1 m %2.%3 s").arg(minutes).arg(seconds).arg(milliseconds, 2, 10, QChar('0'));
     }
 
-    int32_t hours = minutes / 60;
+    int hours = minutes / 60;
     minutes -= 60 * hours;
 
     return QString("%1 h %2 m %3.%4 s").arg(hours).arg(minutes).arg(seconds).arg(milliseconds, 2, 10, QChar('0'));
@@ -57,27 +57,27 @@ QString Frame3::getTime()
 
 QString Frame3::getTimeEgg()
 {
-    int32_t seconds = occidentary / 60;
-    int32_t milliseconds = ((occidentary % 60) * 100) / 60;
+    int seconds = occidentary / 60;
+    int milliseconds = ((occidentary % 60) * 100) / 60;
     if (seconds < 60)
     {
         return QString("%1.%2 s").arg(seconds).arg(milliseconds, 2, 10, QChar('0'));
     }
 
-    int32_t minutes = occidentary / 3600;
+    int minutes = occidentary / 3600;
     seconds -= minutes * 60;
     if (minutes < 60)
     {
         return QString("%1 m %2.%3 s").arg(minutes).arg(seconds).arg(milliseconds, 2, 10, QChar('0'));
     }
 
-    int32_t hours = minutes / 60;
+    int hours = minutes / 60;
     minutes -= 60 * hours;
 
     return QString("%1 h %2 m %3.%4 s").arg(hours).arg(minutes).arg(seconds).arg(milliseconds, 2, 10, QChar('0'));
 }
 
-void Frame3::setInheritanceEmerald(u32 iv1, u32 iv2, u32 par1, u32 par2, u32 par3, u32 inh1, u32 inh2, u32 inh3, QVector<u32> parent1, QVector<u32> parent2)
+void Frame3::setInheritanceEmerald(u16 iv1, u16 iv2, u16 par1, u16 par2, u16 par3, u16 inh1, u16 inh2, u16 inh3, QVector<u8> parent1, QVector<u8> parent2)
 {
     ivs[0] = iv1 & 0x1f;
     ivs[1] = (iv1 >> 5) & 0x1f;
@@ -147,7 +147,7 @@ void Frame3::setInheritanceEmerald(u32 iv1, u32 iv2, u32 par1, u32 par2, u32 par
     power = (30 + ((((ivs[0] >> 1) & 1) + 2 * ((ivs[1] >> 1) & 1) + 4 * ((ivs[2] >> 1) & 1) + 8 * ((ivs[5] >> 1) & 1) + 16 * ((ivs[3] >> 1) & 1) + 32 * ((ivs[4] >> 1) & 1)) * 40 / 63));
 }
 
-void Frame3::setInheritance(u32 iv1, u32 iv2, u32 par1, u32 par2, u32 par3, u32 inh1, u32 inh2, u32 inh3, QVector<u32> parent1, QVector<u32> parent2)
+void Frame3::setInheritance(u16 iv1, u16 iv2, u16 par1, u16 par2, u16 par3, u16 inh1, u16 inh2, u16 inh3, QVector<u8> parent1, QVector<u8> parent2)
 {
     ivs[0] = iv1 & 0x1f;
     ivs[1] = (iv1 >> 5) & 0x1f;
@@ -156,16 +156,16 @@ void Frame3::setInheritance(u32 iv1, u32 iv2, u32 par1, u32 par2, u32 par3, u32 
     ivs[4] = (iv2 >> 10) & 0x1f;
     ivs[5] = iv2 & 0x1f;
 
-    u16 available[6] =  { 0, 1, 2, 3, 4, 5 };
-    u32 val[6] = { inh1, inh2, inh3, par1, par2, par3 };
+    u8 available[6] =  { 0, 1, 2, 3, 4, 5 };
+    u16 val[6] = { inh1, inh2, inh3, par1, par2, par3 };
 
-    for (u32 cnt = 0; cnt < 3; cnt++)
+    for (u8 cnt = 0; cnt < 3; cnt++)
     {
         // Decide which parent (1 or 2) from which we'll pick an IV
-        u32 par = val[3 + cnt] & 1;
+        u8 par = val[3 + cnt] & 1;
 
         // Decide which stat to pick for IV inheritance
-        u32 ivslot = available[val[cnt] % (6 - cnt)];
+        u8 ivslot = available[val[cnt] % (6 - cnt)];
 
         switch (ivslot)
         {
@@ -190,7 +190,7 @@ void Frame3::setInheritance(u32 iv1, u32 iv2, u32 par1, u32 par2, u32 par3, u32 
         }
 
         // Avoids repeat IV inheritance
-        for (int i = ivslot; i < 5; i++)
+        for (u8 i = ivslot; i < 5; i++)
         {
             available[i] = available[i + 1];
         }
