+++ Gen3/Frame3.hpp
@@ -30,10 +30,10 @@ public:
     Frame3(u16 tid, u16 sid, u16 psv);
     QString getTime();
     QString getTimeEgg();
-    void setInheritanceEmerald(u32 iv1, u32 iv2, u32 par1, u32 par2, u32 par3, u32 inh1,
-                               u32 inh2, u32 inh3, QVector<u32> parent1, QVector<u32> parent2);
-    void setInheritance(u32 iv1, u32 iv2, u32 par1, u32 par2, u32 par3, u32 inh1,
-                        u32 inh2, u32 inh3, QVector<u32> parent1, QVector<u32> parent2);
+    void setInheritanceEmerald(u16 iv1, u16 iv2, u16 par1, u16 par2, u16 par3, u16 inh1,
+                               u16 inh2, u16 inh3, QVector<u8> parent1, QVector<u8> parent2);
+    void setInheritance(u16 iv1, u16 iv2, u16 par1, u16 par2, u16 par3, u16 inh1,
+                        u16 inh2, u16 inh3, QVector<u8> parent1, QVector<u8> parent2);
     QString getLockReason() const;
     void setLockReason(const QString &value);
     u32 getOccidentary() const;
