+++ Gen3/Searcher3.hpp
@@ -34,7 +34,7 @@ public:
     Searcher3();
     Searcher3(u16 tid, u16 sid, u32 ratio, const FrameCompare &compare);
     ~Searcher3();
-    QVector<Frame3> search(u32 hp, u32 atk, u32 def, u32 spa, u32 spd, u32 spe);
+    QVector<Frame3> search(u8 hp, u8 atk, u8 def, u8 spa, u8 spd, u8 spe);
     void setup(Method method);
     void setupNatureLock(int num);
     void setEncounter(const EncounterArea3 &value);
@@ -50,13 +50,13 @@ private:
     FrameCompare compare;
     EncounterArea3 encounter;
 
-    QVector<Frame3> searchMethodChannel(u32 hp, u32 atk, u32 def, u32 spa, u32 spd, u32 spe);
-    QVector<Frame3> searchMethodColo(u32 hp, u32 atk, u32 def, u32 spa, u32 spd, u32 spe);
-    QVector<Frame3> searchMethodH124(u32 hp, u32 atk, u32 def, u32 spa, u32 spd, u32 spe);
-    QVector<Frame3> searchMethodXD(u32 hp, u32 atk, u32 def, u32 spa, u32 spd, u32 spe);
-    QVector<Frame3> searchMethodXDColo(u32 hp, u32 atk, u32 def, u32 spa, u32 spd, u32 spe);
-    QVector<Frame3> searchMethod124(u32 hp, u32 atk, u32 def, u32 spa, u32 spd, u32 spe);
-    QVector<Frame3> searchMethod1Reverse(u32 hp, u32 atk, u32 def, u32 spa, u32 spd, u32 spe);
+    QVector<Frame3> searchMethodChannel(u8 hp, u8 atk, u8 def, u8 spa, u8 spd, u8 spe);
+    QVector<Frame3> searchMethodColo(u8 hp, u8 atk, u8 def, u8 spa, u8 spd, u8 spe);
+    QVector<Frame3> searchMethodH124(u8 hp, u8 atk, u8 def, u8 spa, u8 spd, u8 spe);
+    QVector<Frame3> searchMethodXD(u8 hp, u8 atk, u8 def, u8 spa, u8 spd, u8 spe);
+    QVector<Frame3> searchMethodXDColo(u8 hp, u8 atk, u8 def, u8 spa, u8 spd, u8 spe);
+    QVector<Frame3> searchMethod124(u8 hp, u8 atk, u8 def, u8 spa, u8 spd, u8 spe);
+    QVector<Frame3> searchMethod1Reverse(u8 hp, u8 atk, u8 def, u8 spa, u8 spd, u8 spe);
 
 };
